src/requirements.txt
adamp==0.3.0
aiohttp==3.8.1
aiosignal==1.2.0
anykeystore==0.2
apex==0.9.10.dev0
argon2-cffi==21.1.0
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
configparser==5.1.0
cryptacular==1.6.2
cycler==0.10.0
Cython==0.29.24
dataclasses==0.8
datasets==2.4.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.4
docker-pycreds==0.4.0
entrypoints==0.3
filelock==3.3.2
fire==0.4.0
frozenlist==1.2.0
fsspec==2022.1.0
gitdb==4.0.9
GitPython==3.1.18
greenlet==1.1.2
huggingface-hub==0.4.0
hupper==1.10.3
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.1
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
multidict==5.2.0
multiprocess==0.70.12.2
munch==2.5.0
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.5
notebook==6.4.5
numpy==1.19.5
oauthlib==3.1.1
packaging==21.0
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.2
PasteDeploy==2.1.1
pathtools==0.1.2
pbkdf2==1.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plaster==1.0
plaster-pastedeploy==0.7
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.21
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
pyarrow==6.0.1
pycocoevalcap==1.2
pycocotools==2.0.2
pycparser==2.20
Pygments==2.10.0
pyparsing==3.0.3
pyramid==2.0
pyramid-mailer==0.15.1
pyrsistent==0.18.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytorch-warmup==0.0.4
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
regex==2021.10.23
repoze.sendmail==4.4.1
requests==2.26.0
requests-oauthlib==1.3.0
responses==0.17.0
sacremoses==0.0.46
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.8.0
sentence-transformers==2.1.0
sentencepiece==0.1.96
sentry-sdk==1.4.3
shortuuid==1.0.7
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.26
subprocess32==3.5.4
tabulate==0.8.10
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.10.0
torchcontrib==0.0.2
torchtext==0.5.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
transaction==3.0.1
transformers==4.12.2
translationstring==1.4
typing-extensions==3.10.0.2
ujson==4.3.0
urllib3==1.26.7
velruse==1.1.1
venusian==3.0.0
Wand==0.6.10
wandb==0.12.9
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
widgetsnbextension==3.5.1
WTForms==3.0.0
wtforms-recaptcha==0.3.2
xxhash==3.0.0
yarl==1.7.2
yaspin==2.1.0
zipp==3.6.0
zope.deprecation==4.4.0
zope.interface==5.4.0
zope.sqlalchemy==1.6
